refactor: mock app data service
alter the way mock app data service handles seed data to more closely mimic service data cache (async and by flow type).

Additionally includes method to override mock data from other specs that use the mock service (e.g. dynamic-data service). This prevents breakages to other services implementing the mock handler if data changes
