+++ packages/shared/src/utils/async-utils.ts
@@ -1,0 +1,8 @@
+/** helper function used for dev to wait a fixed amount of time */
+export function _wait(ms: number) {
+  return new Promise<void>((resolve) => {
+    setTimeout(() => {
+      resolve();
+    }, ms);
+  });
+}
