+++ packages/shared/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from "./async-utils";
 export * from "./cli-utils";
 export * from "./delimiters";
 export * from "./file-utils";
