+++ src/app/shared/services/data/app-data.service.spec.ts
@@ -3,7 +3,7 @@ import { TestBed } from "@angular/core/testing";
 import { HttpClientTestingModule } from "@angular/common/http/testing";
 
 import { AppDataVariableService } from "./app-data-variable.service";
-import { AppDataService } from "./app-data.service";
+import { AppDataService, IAppDataCache } from "./app-data.service";
 import { FlowTypes } from "../../model";
 import { MockAppDataVariableService } from "./app-data-variable.service.spec";
 import { ErrorHandlerService } from "../error-handler/error-handler.service";
@@ -12,42 +12,57 @@ import { DbService } from "../db/db.service";
 import { MockDbService } from "../db/db.service.spec";
 import { Injectable } from "@angular/core";
 import { ISheetContents } from "src/app/data";
+import { _wait } from "packages/shared/src/utils/async-utils";
 
-const DATA_MOCK: { [flow_name: string]: FlowTypes.FlowTypeWithData } = {
-  flow_a: {
-    flow_name: "flow_a",
-    flow_type: "data_list",
-    rows: [
-      { id: "a_id1", number: 1, string: "a_hello", boolean: true },
-      { id: "a_id2", number: 2, string: "a_goodbye", boolean: false },
-    ],
-  },
-  flow_b: {
-    flow_name: "flow_b",
-    flow_type: "data_list",
-    rows: [
-      { id: "b_id1", number: 1, string: "b_hello", boolean: true },
-      { id: "b_id2", number: 2, string: "b_goodbye", boolean: false },
-    ],
-    _overrides: { flow_c: "true" },
-  },
-  flow_c: {
-    flow_name: "flow_c",
-    flow_type: "data_list",
-    rows: [
-      { id: "c_id1", number: 1, string: "c_hello", boolean: true },
-      { id: "c_id2", number: 2, string: "c_goodbye", boolean: false },
-    ],
-    _overrides: { flow_b: "true" },
-  },
-  flow_d: {
-    flow_name: "flow_d",
-    flow_type: "data_list",
-    rows: [
-      { id: "d_id1", number: 1, string: "d_hello", boolean: true },
-      { id: "d_id2", number: 2, string: "d_goodbye", boolean: false },
-    ],
-    _overrides: { flow_a: "true" },
+/** Base mock data for use with any services calling mock app-data handlers */
+const DATA_CACHE_CLEAN: IAppDataCache = {
+  asset_pack: {},
+  data_list: {},
+  data_pipe: {},
+  generator: {},
+  global: {},
+  template: {},
+  tour: {},
+};
+
+/** Mock data used specifically for the app-data service spec */
+const SPEC_MOCK_DATA: Partial<IAppDataCache> = {
+  data_list: {
+    flow_a: {
+      flow_name: "flow_a",
+      flow_type: "data_list",
+      rows: [
+        { id: "a_id1", number: 1, string: "a_hello", boolean: true },
+        { id: "a_id2", number: 2, string: "a_goodbye", boolean: false },
+      ],
+    },
+    flow_b: {
+      flow_name: "flow_b",
+      flow_type: "data_list",
+      rows: [
+        { id: "b_id1", number: 1, string: "b_hello", boolean: true },
+        { id: "b_id2", number: 2, string: "b_goodbye", boolean: false },
+      ],
+      _overrides: { flow_c: "true" },
+    },
+    flow_c: {
+      flow_name: "flow_c",
+      flow_type: "data_list",
+      rows: [
+        { id: "c_id1", number: 1, string: "c_hello", boolean: true },
+        { id: "c_id2", number: 2, string: "c_goodbye", boolean: false },
+      ],
+      _overrides: { flow_b: "true" },
+    },
+    flow_d: {
+      flow_name: "flow_d",
+      flow_type: "data_list",
+      rows: [
+        { id: "d_id1", number: 1, string: "d_hello", boolean: true },
+        { id: "d_id2", number: 2, string: "d_goodbye", boolean: false },
+      ],
+      _overrides: { flow_a: "true" },
+    },
   },
 };
 
@@ -83,12 +98,18 @@ const CONTENTS_MOCK: ISheetContents = {
 
 /** Mock calls for sheets from the appData service to return test data */
 export class MockAppDataService implements Partial<AppDataService> {
+  public appDataCache: IAppDataCache;
+
+  // allow additional specs implementing service to provide their own data if required
+  constructor(mockData: Partial<IAppDataCache> = {}) {
+    this.appDataCache = { ...DATA_CACHE_CLEAN, ...mockData };
+  }
   public async getSheet<T extends FlowTypes.FlowTypeWithData>(
     flow_type: FlowTypes.FlowType,
     flow_name: string
   ): Promise<T> {
-    const rows = DATA_MOCK[flow_name].rows || [];
-    return { flow_name, flow_type, rows } as any;
+    await _wait(50);
+    return this.appDataCache[flow_type]?.[flow_name] as T;
   }
 }
 
@@ -97,15 +118,7 @@ export class MockAppDataService implements Partial<AppDataService> {
 class AppDataServiceExtended extends AppDataService {
   protected sheetContents = CONTENTS_MOCK;
   protected translationContents = {};
-  public appDataCache = {
-    asset_pack: {},
-    data_list: { ...DATA_MOCK },
-    data_pipe: {},
-    generator: {},
-    global: {},
-    template: {},
-    tour: {},
-  };
+  public appDataCache = { ...DATA_CACHE_CLEAN, ...SPEC_MOCK_DATA };
 }
 
 /********************************************************************************
